night/backend/requirements.txt
# FastAPI Stack
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
python-multipart==0.0.9

# ML
xgboost==2.0.3
shap==0.45.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.0

# AI
google-generativeai==0.5.4

# Redis
redis==5.0.4
hiredis==2.3.2

# HTTP
httpx==0.27.0
aiohttp==3.9.5

# Utils
python-dateutil==2.9.0
structlog==24.1.0

# Async SQLite (dev)
aiosqlite==0.20.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
